Library_Resources/requirements.txt
python-docx==0.8.11
markdownify==0.11.6
html2text==2020.1.16
